Further directory and module refactoring
