+++ kptracker_serv/kpbt/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-from .models import User
-from django.contrib.auth.admin import UserAdmin
+
 
 # Register your models here.
