+++ kptracker_serv/kpbt/bowling_centers/views.py
@@ -1,6 +1,5 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
-from kpbt import models
 
 def create_bowling_center(request):
 	return render(request, 'bowling_centers/create.html', context={})
