--- kptracker_serv/kpbt/serializers.py
@@ -1,6 +1,0 @@
-from rest_framework import serializers
-
-#Define serializers here
-
-
-
